pyproject.toml
[tool.poetry]
name = "ai-scheduler"
version = "0.1.0"
description = "AI予定管理秘書アプリ - LINEでGoogleカレンダー連携"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "app"}]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.105.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-dotenv = "^1.0.0"
line-bot-sdk = "^3.5.0"
google-api-python-client = "^2.108.0"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.1.0"
langchain = "^0.0.335"
google-generativeai = "^0.3.1"
firebase-admin = "^6.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
